+++ src/Arius.Cli.Tests/CliTests.cs
@@ -1,5 +1,6 @@
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Features.ChunkHydrationStatusQuery;
+using Arius.Core.Features.ListQuery;
 using Arius.Core.Features.Restore;
 using Arius.Core.Shared.Storage;
 using Mediator;
@@ -7,7 +8,6 @@ using Microsoft.Extensions.DependencyInjection;
 using NSubstitute;
 using Shouldly;
 using System.CommandLine;
-using Arius.Core.Features.ListQuery;
 
 namespace Arius.Cli.Tests;
