+++ src/Arius.Cli.Tests/ProgressTests.cs
@@ -1,7 +1,7 @@
 using Arius.Cli.Commands.Archive;
 using Arius.Cli.Commands.Restore;
 using Arius.Core;
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Features.Restore;
 using Arius.Core.Shared.Storage;
 using Mediator;
