+++ src/Arius.Cli/Commands/Archive/ArchiveVerb.cs
@@ -1,5 +1,5 @@
 using Arius.AzureBlob;
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Shared.Storage;
 using Humanizer;
 using Mediator;
