+++ src/Arius.Cli/Commands/Ls/LsVerb.cs
@@ -1,4 +1,5 @@
 using Arius.AzureBlob;
+using Arius.Core.Features.ListQuery;
 using Arius.Core.Shared.Storage;
 using Humanizer;
 using Mediator;
@@ -6,7 +7,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Serilog;
 using Spectre.Console;
 using System.CommandLine;
-using Arius.Core.Features.ListQuery;
 
 namespace Arius.Cli.Commands.Ls;
