+++ src/Arius.Core/Features/ContainerNamesQuery/ContainerNamesQuery.cs
@@ -1,7 +1,7 @@
-using System.Runtime.CompilerServices;
 using Arius.Core.Shared.Storage;
 using Mediator;
 using Microsoft.Extensions.DependencyInjection;
+using System.Runtime.CompilerServices;
 
 namespace Arius.Core.Features.ContainerNamesQuery;
