+++ src/Arius.Core/Features/ListQuery/ListQueryHandler.cs
@@ -1,4 +1,3 @@
-using System.Runtime.CompilerServices;
 using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.FileTree;
@@ -6,6 +5,7 @@ using Arius.Core.Shared.Snapshot;
 using Arius.Core.Shared.Storage;
 using Mediator;
 using Microsoft.Extensions.Logging;
+using System.Runtime.CompilerServices;
 
 namespace Arius.Core.Features.ListQuery;
