+++ src/Arius.Core/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Features.ChunkHydrationStatusQuery;
 using Arius.Core.Features.ContainerNamesQuery;
 using Arius.Core.Features.ListQuery;
