+++ src/Arius.E2E.Tests/E2ETests.cs
@@ -1,5 +1,5 @@
 using Arius.AzureBlob;
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Features.Restore;
 using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Encryption;
