+++ src/Arius.Explorer/ChooseRepository/ChooseRepositoryViewModel.cs
@@ -1,3 +1,4 @@
+using Arius.Core.Features.ContainerNamesQuery;
 using Arius.Explorer.Settings;
 using Arius.Explorer.Shared.Extensions;
 using CommunityToolkit.Mvvm.ComponentModel;
@@ -13,7 +14,6 @@ using System.Reactive.Subjects;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
-using Arius.Core.Features.ContainerNamesQuery;
 using Unit = System.Reactive.Unit;
 
 namespace Arius.Explorer.ChooseRepository;
