+++ src/Arius.Explorer/RepositoryExplorer/FileItemViewModel.cs
@@ -1,8 +1,8 @@
 using Arius.Core.Features.ChunkHydrationStatusQuery;
+using Arius.Core.Features.ListQuery;
 using CommunityToolkit.Mvvm.ComponentModel;
 using System.IO;
 using System.Windows.Media;
-using Arius.Core.Features.ListQuery;
 
 namespace Arius.Explorer.RepositoryExplorer;
