+++ src/Arius.Explorer/RepositoryExplorer/RepositoryExplorerViewModel.cs
@@ -1,4 +1,5 @@
 using Arius.Core.Features.ChunkHydrationStatusQuery;
+using Arius.Core.Features.ListQuery;
 using Arius.Core.Features.Restore;
 using Arius.Explorer.Infrastructure;
 using Arius.Explorer.Settings;
@@ -16,7 +17,6 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
-using Arius.Core.Features.ListQuery;
 
 namespace Arius.Explorer.RepositoryExplorer;
