+++ src/Arius.Integration.Tests/Pipeline/ContainerCreationTests.cs
@@ -1,5 +1,5 @@
 using Arius.AzureBlob;
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.Storage;
