+++ src/Arius.Integration.Tests/Pipeline/RestorePointerTimestampTests.cs
@@ -1,4 +1,4 @@
-using Arius.Core.Features.Archive;
+using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Features.Restore;
 using Arius.Core.Shared.Storage;
 using Arius.Integration.Tests.Storage;
